backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asgiref==3.8.1
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cramjam==2.10.0
cycler==0.12.1
distro==1.9.0
fastparquet==2024.11.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.0
kiwisolver==1.4.8
llvmlite==0.45.0
MarkupSafe==3.0.2
matplotlib==3.10.3
nfl_data_py==0.3.3
numba==0.62.0
numpy==1.26.4
openai==1.108.0
packaging==25.0
pandas==1.5.3
pillow==11.2.1
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.3.1
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
